pyproject.toml
[project]
name = "alice-rag-llm"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Philippe Miron", email = "philippemiron@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.4",
    "llama-index>=0.12.30",
    "llama-index-embeddings-huggingface>=0.5.3",
    "pyyaml>=6.0.2",
    "streamlit>=1.44.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "sphinx>=8.1.3",
    "sphinx-book-theme>=1.1.4",
    "sphinx-copybutton>=0.5.2",
]

# clean: rm -Rf docs/_build/ docs/_autosummary/ src/store src/documents
# docs: uv run sphinx-build -M html docs docs/_build -W --keep-going
# embeddings: uv run python src/store.py
# uv run pytest tests